remove code duplication by moving MyLSPScorer one level up.
